feat(task): points 支持 0.5-5 小数（0.5 步长）
- V6 迁移：tasks.points 改 numeric(2,1)，存量 >5 归一化为 5、<=0 置 NULL（先归一化再 ALTER，避免溢出）
- Task.points 改 BigDecimal；REST 创建/PATCH 与 MCP create_tasks 统一校验：
  非空时 0.5<=p<=5 且为 0.5 倍数，否则 400 INVALID_POINTS（中文文案，agent 可读）
- 波及求和全部 BigDecimal：CapacityService.assignedPoints、BurndownService remaining、
  Roadmap donePoints/totalPoints、Dashboard donePct；JSON 仍输出数字
- MCP create_tasks 的 points schema 改 number（minimum 0.5 / maximum 5 / multipleOf 0.5）

+++ backend/src/main/java/pm/mcp/McpTools.java
@@ -19,6 +19,7 @@ import pm.task.TaskService;
 import pm.tenant.TenantContext;
 import pm.tenant.TenantEntity;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
@@ -62,7 +63,7 @@ public class McpTools {
     public record ProjectItem(String key, String name) {
     }
 
-    public record TaskInput(Task.Type type, String title, String description, Integer points, Long epicId) {
+    public record TaskInput(Task.Type type, String title, String description, BigDecimal points, Long epicId) {
     }
 
     public record CreatedTask(String seq, String title) {
