+++ backend/src/main/java/pm/task/TaskBrief.java
@@ -1,10 +1,12 @@
 package pm.task;
 
+import java.math.BigDecimal;
+
 /**
  * 任务摘要视图：roadmap / dashboard / all-sprints 等列表复用。
  */
 public record TaskBrief(Long id, int seq, String title, Task.Type type, Task.Status status,
-                        Integer points, Long assigneeId) {
+                        BigDecimal points, Long assigneeId) {
 
     public static TaskBrief from(Task t) {
         return new TaskBrief(t.getId(), t.getSeq(), t.getTitle(), t.getType(), t.getStatus(),
